# gin-blog-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/config": {
            "get": {
                "description": "获取配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog_info"
                ],
                "summary": "获取配置信息",
                "parameters": [
                    {
                        "description": "配置信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-map_string_string"
                        }
                    }
                }
            },
            "patch": {
                "description": "更新配置信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog_info"
                ],
                "summary": "更新配置信息",
                "parameters": [
                    {
                        "description": "更新配置信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-any"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handle.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-handle_LoginVO"
                        }
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "description": "退出登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "退出登录",
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserAuth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handle.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/report": {
            "post": {
                "description": "用户登进后台时上报信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blog_info"
                ],
                "summary": "上报用户信息",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/handle.Response-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handle.LoginReq": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handle.LoginVO": {
            "type": "object",
            "properties": {
                "article_like_set": {
                    "description": "点赞 Set： 用于记录用户点赞过的文章，评论",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "avatar": {
                    "description": "用户头像，最大长度1024字符，不能为空",
                    "type": "string"
                },
                "comment_like_set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "description": "用户的邮箱，最大长度30字符，保存用户的电子邮件地址",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "intro": {
                    "description": "用户个人简介，最大长度255字符，用于描述用户的个人信息或介绍",
                    "type": "string"
                },
                "nickname": {
                    "description": "用户的昵称，唯一，最大长度30字符，不能为空",
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "website": {
                    "description": "用户的个人网站链接，最大长度255字符，用于存储用户的官网、博客等链接",
                    "type": "string"
                }
            }
        },
        "handle.RegisterReq": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 4
                }
            }
        },
        "handle.Response-any": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handle.Response-handle_LoginVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/handle.LoginVO"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "handle.Response-map_string_string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/map_string_string"
                        }
                    ]
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
        "map_string_string": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        }
    }
}